requirements.txt
fastapi
uvicorn[standard]
motor
websockets
python-dotenv
pymongo[srv]
dnspython
pydantic-settings
groq
twilio
graphiti-core[google-genai]